Bug 57480: Fix some unnecessary casts, generics, Eclipse warnings, ...
git-svn-id: https://svn.apache.org/repos/asf/poi/trunk@1658594 13f79535-47bb-0310-9956-ffa450edef68

+++ src/java/org/apache/poi/hssf/dev/RecordLister.java
@@ -154,7 +154,7 @@ public class RecordLister
        return s.toString();
     }
     private static String byteToHex(byte b) {
-       int i = (int)b;
+       int i = b;
        if(i<0) {
           i += 256;
        }
